+++ src/IRule.php
@@ -2,7 +2,12 @@
 
 namespace Saritasa\Laravel\Validation;
 
+/**
+ * Each rule or rule set in this package inherits this interface,
+ * so we can distinguish them from other classes
+ */
 interface IRule
 {
+    /** Magic method, called on explicit or implicit casting to string */
     function __toString(): string;
 }