+++ src/IntRuleSet.php
@@ -2,6 +2,9 @@
 
 namespace Saritasa\Laravel\Validation;
 
+/**
+ * Validation rules for integer type values
+ */
 class IntRuleSet extends NumericRuleSet
 {
     protected $primaryRuleName = 'integer';