+++ src/Rules/Exists.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Rules;
-
-class Exists
-{
-
-}