+++ CHANGES.md
@@ -1,8 +1,15 @@
 # Changes History
 
+1.0.3
+-----
+Add ServiceProvider - allows to use rule builder without explicit ->toString()
+
+1.0.1, 1.0.2
+-----
+Add more rules, tests, and inline documentation
+
 1.0.0
 -----
-
 Basic rules:
 - required
 - required_with